requirements.txt
pywinauto
pygetwindow
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
pyautogui
gTTS